+++ Documentation/RxGRDBDemo/RxGRDBDemo/PlaceAnnotation.swift
@@ -17,6 +17,10 @@ final class PlaceAnnotation: NSObject, MKAnnotation {
     init(place: Place) {
         self.place = place
     }
+    
+    func update(from row: Row) {
+        self.place = Place(row: row)
+    }
 }
 
 // Have PlaceAnnotation adopt RowConvertible, Persistable, and Diffable, so
@@ -36,12 +40,3 @@ extension PlaceAnnotation: Persistable {
     }
 }
 
-extension PlaceAnnotation: Diffable {
-    // The `primaryKeySortedDiff` observable documents that this method should
-    // return self if we wish elements to be reused during the computation of
-    // the diff.
-    func updated(with row: Row) -> Self {
-        place = Place(row: row)
-        return self
-    }
-}