+++ RxGRDB/TypedRequest+Rx.swift
@@ -14,31 +14,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: RowConvertible {
     /// committed database transaction that has modified the tables and columns
     /// fetched by the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first array
-    /// is emitted synchronously, on subscription.
+    /// All arrays are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchAll(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<[Base.RowDecoder]>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try Row.fetchAll($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try Row.fetchAll($0, request) }
                 .distinctUntilChanged(==)
                 .map { (rows: [Row]) in rows.map { Base.RowDecoder.init(row: $0) } }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchAll($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchAll($0) }
         }
     }
     
@@ -46,31 +48,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: RowConvertible {
     /// database transaction that has modified the tables and columns fetched by
     /// the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first record
-    /// is emitted synchronously, on subscription.
+    /// All records are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchOne(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<Base.RowDecoder?>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try Row.fetchOne($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try Row.fetchOne($0, request) }
                 .distinctUntilChanged(==)
                 .map { (row: Row?) in row.map { Base.RowDecoder.init(row: $0) } }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchOne($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchOne($0) }
         }
     }
 }
@@ -82,30 +86,32 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: Row {
     /// committed database transaction that has modified the tables and columns
     /// fetched by the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first array
-    /// is emitted synchronously, on subscription.
+    /// All arrays are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchAll(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<[Row]>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchAll($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchAll($0) }
                 .distinctUntilChanged(==)
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchAll($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchAll($0) }
         }
     }
     
@@ -113,30 +119,32 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: Row {
     /// database transaction that has modified the tables and columns fetched by
     /// the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first row
-    /// is emitted synchronously, on subscription.
+    /// All rows are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchOne(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<Row?>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchOne($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchOne($0) }
                 .distinctUntilChanged(==)
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchOne($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchOne($0) }
         }
         
     }
@@ -149,31 +157,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: DatabaseValueConve
     /// committed database transaction that has modified the tables and columns
     /// fetched by the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first array
-    /// is emitted synchronously, on subscription.
+    /// All arrays are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchAll(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<[Base.RowDecoder]>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try DatabaseValue.fetchAll($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try DatabaseValue.fetchAll($0, request) }
                 .distinctUntilChanged(==)
                 .map { (dbValues: [DatabaseValue]) in dbValues.map { $0.losslessConvert() as Base.RowDecoder } }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchAll($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchAll($0) }
         }
     }
     
@@ -181,31 +191,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: DatabaseValueConve
     /// database transaction that has modified the tables and columns fetched by
     /// the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first value
-    /// is emitted synchronously, on subscription.
+    /// All values are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchOne(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<Base.RowDecoder?>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try DatabaseValue.fetchOne($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try DatabaseValue.fetchOne($0, request) }
                 .distinctUntilChanged(==)
                 .map { (dbValue: DatabaseValue?) in dbValue.map { $0.losslessConvert() as Base.RowDecoder? } ?? nil }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchOne($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchOne($0) }
         }
     }
 }
@@ -225,31 +237,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: _OptionalProtocol,
     /// committed database transaction that has modified the tables and columns
     /// fetched by the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first array
-    /// is emitted synchronously, on subscription.
+    /// All arrays are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchAll(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<[Base.RowDecoder._Wrapped?]>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try DatabaseValue.fetchAll($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try DatabaseValue.fetchAll($0, request) }
                 .distinctUntilChanged(==)
                 .map { (dbValues: [DatabaseValue]) in dbValues.map { $0.losslessConvert() as Base.RowDecoder._Wrapped? } }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try Optional<Base.RowDecoder._Wrapped>.fetchAll($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try Optional<Base.RowDecoder._Wrapped>.fetchAll($0, request) }
         }
     }
 }
@@ -261,31 +275,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: DatabaseValueConve
     /// committed database transaction that has modified the tables and columns
     /// fetched by the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first array
-    /// is emitted synchronously, on subscription.
+    /// All arrays are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchAll(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<[Base.RowDecoder]>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try DatabaseValue.fetchAll($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try DatabaseValue.fetchAll($0, request) }
                 .distinctUntilChanged(==)
                 .map { (dbValues: [DatabaseValue]) in dbValues.map { $0.losslessConvert() as Base.RowDecoder } }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchAll($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchAll($0) }
         }
     }
     
@@ -293,31 +309,33 @@ extension Reactive where Base: TypedRequest, Base.RowDecoder: DatabaseValueConve
     /// database transaction that has modified the tables and columns fetched by
     /// the request.
     ///
-    /// If you set `synchronizedStart` to true (the default), the first value
-    /// is emitted synchronously, on subscription.
+    /// All values are emitted on *scheduler*, which defaults to
+    /// `MainScheduler.instance`. If you set *synchronizedStart* to true (the
+    /// default value), the first element is emitted right upon subscription.
     ///
     /// - parameter writer: A DatabaseWriter (DatabaseQueue or DatabasePool).
     /// - parameter synchronizedStart: When true (the default), the first
     ///   element is emitted synchronously, on subscription.
-    /// - parameter resultQueue: A DispatchQueue (default is the main queue).
+    /// - parameter scheduler: The scheduler on which elements are emitted
+    ///   (default is MainScheduler.instance).
     public func fetchOne(
         in writer: DatabaseWriter,
         synchronizedStart: Bool = true,
-        distinctUntilChanged: Bool = false,
-        resultQueue: DispatchQueue = DispatchQueue.main)
+        scheduler: SerialDispatchQueueScheduler = MainScheduler.instance,
+        distinctUntilChanged: Bool = false)
         -> Observable<Base.RowDecoder?>
     {
         let request = base
         if distinctUntilChanged {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try DatabaseValue.fetchOne($0, request) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try DatabaseValue.fetchOne($0, request) }
                 .distinctUntilChanged(==)
                 .map { (dbValue: DatabaseValue?) in dbValue.map { $0.losslessConvert() as Base.RowDecoder? } ?? nil }
         } else {
             return AnyDatabaseWriter(writer).rx
-                .changeTokens(in: [request], synchronizedStart: synchronizedStart)
-                .mapFetch(resultQueue: resultQueue) { try request.fetchOne($0) }
+                .changeTokens(in: [request], synchronizedStart: synchronizedStart, scheduler: scheduler)
+                .mapFetch { try request.fetchOne($0) }
         }
     }
 }